requirements.txt
matplotlib~=3.10.3
numpy~=2.2.5
casadi~=3.7.0
PyQt5
gym~=0.26.2
torch~=2.7.0
rich~=14.0.0
imageio~=2.37.0
pillow~=11.2.1
pandas~=2.2.3
wandb~=0.19.11
tabulate~=0.9.0